Keys & Verbosity
Initialize now takes in a verbosity flag, and many of the unnecessary prints are hidden behind it.
Additionally machinery is now in place for the C code to generate a private key which can then be passed into initialize to maintain a node's identity! (if null is passed in a new identity is automatically generated!)