Windows (MSYS) Build Support
Tweaked the build system so that when MSYS is installed you can build on windows! Everything worked except that the automatically detected GOROOT was wrong, so a method to tweak it has been added and its value is now being set based on the location of the compiler!

(There has also been tentative success compiling the C API with Visual Studio!)