Multiple Networks
Update the GO code to support having more than one network in an application!
This means that every call in the C api now needs the network it should work with.
The C++ API has had all of the singleton functions removed.
Callbacks in both APIs now provide the network that called them as a parameter (thus the same callback code can be used with multiple networks!)